separate out gaussian splats from GLTF exporter as well into an extension.
